Fix sparse_set to allow large domains.

+++ src/domain/sparse_set.rs
@@ -6,21 +6,21 @@ use std::fmt::Display;
 #[doc(hidden)]
 #[derive(Clone, Debug, PartialEq)]
 pub struct SparseSetState {
-    pub size: u16,
-    pub min: u16,
-    pub max: u16,
+    pub size: u32,
+    pub min: u32,
+    pub max: u32,
 }
 
 #[doc(hidden)]
 #[derive(Debug, Clone)]
 pub struct SparseSet {
     off: i32,      // the domain offset (fixed)
-    n: u16,        // total number of values in domain
-    min: u16,      // current minimum value in the set
-    max: u16,      // current maximum value in the set
-    size: u16,     // domain size
-    ind: Vec<u16>, // sparse set indices
-    val: Vec<u16>, // sparse set values
+    n: u32,        // total number of values in domain
+    min: u32,      // current minimum value in the set
+    max: u32,      // current maximum value in the set
+    size: u32,     // domain size
+    ind: Vec<u32>, // sparse set indices
+    val: Vec<u32>, // sparse set values
 }
 
 impl Display for SparseSet {
@@ -55,11 +55,11 @@ impl SparseSet {
         }
         
         let maxmin = (max - min) as u32;
-        let n = (maxmin + 1) as u16;
+        let n = (maxmin + 1) as u32;
         SparseSet {
             off: min,
             min: 0,
-            max: maxmin as u16,
+            max: maxmin,
             n,
             size: n,
             ind: (0..n).collect(),
@@ -129,17 +129,17 @@ impl SparseSet {
             return false;
         }
         let v = v - self.off;
-        self.contains_intl(v as u16)
+        self.contains_intl(v as u32)
     }
     // This method operates on the shifted value (one cannot shift now).
-    fn contains_intl(&self, v: u16) -> bool {
+    fn contains_intl(&self, v: u32) -> bool {
         if v >= self.n {
             false
         } else {
             self.ind[v as usize] < self.size
         }
     }
-    pub fn exchange(&mut self, val1: u16, val2: u16) {
+    pub fn exchange(&mut self, val1: u32, val2: u32) {
         let v1 = val1;
         let v2 = val2;
         let i1 = self.ind[v1 as usize];
@@ -150,12 +150,12 @@ impl SparseSet {
         self.ind[v2 as usize] = i1;
     }
 
-    fn update_bounds_val_removed(&mut self, val: u16) {
+    fn update_bounds_val_removed(&mut self, val: u32) {
         self.update_max_val_removed(val);
         self.update_min_val_removed(val);
     }
     // update after max value is removed
-    fn update_max_val_removed(&mut self, val: u16) {
+    fn update_max_val_removed(&mut self, val: u32) {
         if !self.is_empty() && self.max == val {
             // The maximum was removed, search the new one
             for v in (self.min..val).rev() {
@@ -167,7 +167,7 @@ impl SparseSet {
         }
     }
     // update after min value is removed
-    fn update_min_val_removed(&mut self, val: u16) {
+    fn update_min_val_removed(&mut self, val: u32) {
         if !self.is_empty() && self.min == val {
             // The minimum was removed, search the new one
             let vv = val + 1;
@@ -186,7 +186,7 @@ impl SparseSet {
             // The value has already been removed
             return false;
         }
-        let val = (val - self.off) as u16;
+        let val = (val - self.off) as u32;
         self.exchange(val, self.val[self.size() - 1]);
         self.size = self.size - 1;
         self.update_bounds_val_removed(val);
@@ -205,7 +205,7 @@ impl SparseSet {
             self.remove_all();
             return;
         }
-        let v = (v - self.off) as u16;
+        let v = (v - self.off) as u32;
         let val = self.val[0];
         let index = self.ind[v as usize];
         self.ind[v as usize] = 0;
@@ -312,7 +312,7 @@ impl SparseSet {
             // Check if value is in our universe
             if val >= self.off && val < self.off + self.n as i32 {
                 // Manually add the value (similar to remove but in reverse)
-                let val_internal = (val - self.off) as u16;
+                let val_internal = (val - self.off) as u32;
                 if !self.contains_intl(val_internal) {
                     // Add value to the end of the active set
                     let new_pos = self.size;
@@ -382,13 +382,13 @@ impl SparseSet {
     }
     
     /// Get current state size (for simple size-only backtracking)
-    pub fn current_size(&self) -> u16 {
+    pub fn current_size(&self) -> u32 {
         self.size
     }
     
     /// Restore to a specific size (assumes min/max are still valid)
     /// WARNING: Only use if you're certain min/max haven't changed since the size was recorded
-    pub fn restore_size(&mut self, size: u16) {
+    pub fn restore_size(&mut self, size: u32) {
         debug_assert!(size <= self.n, "Cannot restore to size larger than universe");
         self.size = size;
     }
